3/requirements.txt
# Core AI Agent Dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0

# MCP Client Dependencies
mcp>=1.0.0
httpx>=0.27.0
websockets>=15.0

# Environment and Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.5.0

# Logging and CLI
structlog>=24.1.0
rich>=13.7.0

# JSON-RPC client for MCP
jsonrpc-websocket>=3.1.0